refactor(canvas): extract Konva shape components to canvas/ui modules (#192)
Moves all Konva object renderers and UI components out of the
3,698-line monolith into dedicated modules:

- ObjectShapes.tsx  — GridLines, RoadSegment, PolylineRoad, CurveRoad,
                      CubicBezierRoad, SignShape, DeviceShape, WorkZone,
                      ArrowShape, TextLabel, MeasurementShape, TaperShape,
                      ObjectShape (dispatch)
- DrawingOverlays.tsx — in-progress draw previews + snap indicator
- ToolButton.tsx    — toolbar button component
- SignEditorPanel.tsx — custom sign editor with color/shape/text controls

traffic-control-planner.tsx: 2,576 → 1,712 lines (-33%)
